third_party/blink/web_tests/accessibility/selection-change-notification-input.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<input autofocus id="input" value="input">

<script>
'use strict';

async_test((t) => {
  // This forces building the accessibility tree, because selection change
  // events only fire on elements that already exist.
  accessibilityController.accessibleElementById('dummy');

  const inputElement = document.getElementById('input');
  const axRootElement = accessibilityController.rootElement;
  const axInputElement = accessibilityController.accessibleElementById('input');

  let gotDocumentSelectionChanged = false;
  let gotSelectedTextChanged = false;
  const succeedIfDone = t.step_func(() => {
    if (gotDocumentSelectionChanged && gotSelectedTextChanged)
      t.done();
  });

  axRootElement.setNotificationListener(function(notification, intents) {
    if (notification == 'DocumentSelectionChanged') {
      assert_array_equals(intents,
                          ['AXEventIntent(setSelection,character,forward)']);
      gotDocumentSelectionChanged = true;
      axRootElement.unsetNotificationListener();
      succeedIfDone();
    }
  });

  axInputElement.setNotificationListener(function(notification, intents) {
    if (notification == 'SelectedTextChanged') {
      assert_array_equals(intents,
                          ['AXEventIntent(setSelection,character,forward)']);
      gotSelectedTextChanged = true;
      axInputElement.unsetNotificationListener();
      succeedIfDone();
    }
  });

  inputElement.setSelectionRange(0, 1);
});

</script>
